Increasing precisement
- Precisement increasing due to the latest refactoring.

(I was assigning spheres to limbs by checking their distance to the main limbs and take the closest distance to find the closest limb. This algorithm works perfectly when it's working with a detailed limb list.

But when I used this algorithm to find the closest main limb, it wasn't working right. Because closest main limbs were sometimes not real parent limbs. For example, spheres on hand should be the parent of the shoulder, but actually, it's closer to the leg so the algorithm was working like "spheres at hand is a child of the leg".

I fixed that bug and increased precisement on finding main limbs.